backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
rapidfuzz==3.11.0
requests==2.32.3
